+++ HISTORY.md
@@ -1,3 +1,8 @@
+## 2012-08-07 ##
+- Added client-install npm script to begin streamlining client-side install
+- Added install and update npm scripts to provide a notification on how to install luhn for client-side use.
+- package version -> 1.0.5
+
 ## 2012-07-02 ##
 - Updated package to use `keywords` instead of `tags`
 - package version -> 1.0.4